PythonCode/requirements.txt
Flask
flask-cors
textblob
google-generativeai
python-dotenv
sentence-transformers
langchain
langchain-community
langchain-groq
chromadb
langchain-huggingface
